+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 ## 0.9.1
 
 ### Added
+
+- `Transform(Plane plane)`
+- `Polygon.Trimmed(Plane plane, bool flip)`
 - ~~`GetSolid()` method on GeometricElement that returns the Csg solid.~~
 - `ToMesh()` method on GeometricElement that return the mesh of a processed solid.
 - `Polygon.ToTransform()`
 - `Elements.Anaysis.AnalysisImage`
-
-### Removed
+- `Profile.CreateFromPolygons(IEnumerable<Polygon> polygons)`
 
 ### Changed